Downgrade nightly version for rustfmt if necessary
Sometimes, not all components are available in any given nightly, which
causes errors such as https://github.com/ilai-deutel/kibi/issues/24.
Therefore, we use the `--allow-downgrade` flag to downgrade the version
if rustfmt is not available for this or any newer nightly.

In the future, `rustup component add` will have a `--allow-downgrade`
flag (see https://github.com/rust-lang/rustup/issues/2146) we can use
instead of this work-around.