Add and update some syntaxes (nu, lrc, rss) (#433)
* Update and expand nu syntax file

Nu is constantly changing, so some commands are gone and others appeared.

Defining the things with `-` in their name (`is-admin`, `not-in`, etc)
seems to break something, so i left them out.

* Add lrc syntax

<https://en.wikipedia.org/wiki/LRC_(file_format)>

Since normal timestamps are fairly easy to spot and have syntax
within them i decided to use the multiline_comment_delims for A2 data.

* Add rss syntax

Yes it is technically xml, but it has its own extension and
different keywords.